% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_tinystan_path}
\alias{get_tinystan_path}
\title{Get the path to TinyStan.}
\usage{
get_tinystan_path()
}
\description{
By default this is set to the value of the environment
variable \code{TINYSTAN}.
}
\details{
If there is no path set, this function will download
a matching version of TinyStan to a folder called
\code{.tinystan} in the user's home directory.
}
\seealso{
\link{set_tinystan_path}
}
