% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{HMCMetric}
\alias{HMCMetric}
\title{Enumeration \code{HMCMetric}}
\format{
An object of class \code{list} of length 3.
}
\usage{
HMCMetric
}
\description{
Choices of metric for HMC. Can select from \verb{$UNIT}, \verb{$DENSE}, and \verb{$DIAGONAL}.
}
\keyword{datasets}
