% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinystan.R
\name{laplace_sampler.tinystan_model}
\alias{laplace_sampler.tinystan_model}
\title{Run Stan's Laplace approximation algorithm}
\usage{
\method{laplace_sampler}{tinystan_model}(
  modal,
  mode,
  data = "",
  num_draws = 1000,
  jacobian = TRUE,
  calculate_lp = TRUE,
  save_hessian = FALSE,
  seed = NULL,
  refresh = 0,
  num_threads = -1
)
}
\description{
Run Stan's Laplace approximation algorithm
}
