% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{OptimizationAlgorithm}
\alias{OptimizationAlgorithm}
\title{Enumeration \code{OptimizationAlgorithm}}
\format{
An object of class \code{list} of length 3.
}
\usage{
OptimizationAlgorithm
}
\description{
Choices of optimization algorithm. Can select from \verb{$NEWTON}, \verb{$BFGS}, and \verb{$LBFGS}.
}
\keyword{datasets}
