% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{set_tinystan_path}
\alias{set_tinystan_path}
\title{Function \code{set_tinystan_path()}}
\usage{
set_tinystan_path(path)
}
\description{
Set the path to TinyStan.
}
\details{
This should point to the top-level folder of the repository.
}
