% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinystan-package.R
\docType{package}
\name{tinystan-package}
\alias{tinystan-package}
\alias{tinystan}
\title{tinystan: A tiny interface to Stan algorithms}
\description{
\if{html}{
   \figure{logo.png}{options: width='50' alt='https://mc-stan.org/about/logo/'}
   \emph{Stan Development Team}
}

\pkg{tinystan}: A tiny \R interface to Stan algorithms.
}
\details{
tinystan (\pkg{tinystan} package) is an interface to Stan
(\href{https://mc-stan.org}{mc-stan.org}) for \R users. It provides the
necessary objects and functions to compile a Stan program and run Stan's
algorithms from \R.
}
\author{
\strong{Maintainer}: Brian Ward \email{bward@flatironinstitute.org}

}
