% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinystan.R
\name{tinystan_model}
\alias{tinystan_model}
\title{Class \code{tinystan_model}}
\usage{
tinystan_model(lib, stanc_args = NULL, make_args = NULL, warn = TRUE)
}
\description{
Load a Stan model, compiling if necessary.
}
\examples{
data_file <- system.file('bernoulli.data.json', package = 'tinystan')
mod <- tinystan_model(system.file('bernoulli.stan', package = 'tinystan'))
fit = sampler(model = mod, data = data_file)
fit

}
