% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffistan.R
\name{handle_error}
\alias{handle_error}
\title{Get and free the error message stored at the C++ pointer}
\usage{
handle_error(rc, lib_name, err_ptr)
}
\description{
Get and free the error message stored at the C++ pointer
}
\keyword{internal}
